+++ src/pages/login.tsx
@@ -17,18 +17,6 @@ export default function LoginForm() {
             email,
             password,
         })
-        const response = await fetch('/api/getCategories', {
-            method: 'GET'
-        })
-        const data = await response.json()
-        console.log(data);
-
-        const Categories = data.categories.map((cat: { name: string; merchants: { name: string }[] }) => {
-            const merchantNames = cat.merchants.map(m => m.name);
-            return [cat.name, ...merchantNames];
-        });
-        
-        localStorage.setItem('categories', JSON.stringify(Categories))
 
         if (res?.ok) {
             router.push('/')
